% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-variables.R
\name{make_acs_variables}
\alias{make_acs_variables}
\title{Prepare ACS Data Variables}
\usage{
make_acs_variables(acs_data, acs_tables)
}
\arguments{
\item{acs_data}{Tibble, description.}

\item{acs_tables}{desc}
}
\value{
a \code{tibble}
}
\description{
Return a \code{tibble} of all of the American Community Survey data variables.
}
