% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acs_tables}
\alias{acs_tables}
\title{A Tibble of ACS Tables}
\format{A \code{tibble} with four (4) rows and three (3) columns.

The object includes the following variables:
\describe{

\item{\code{• NAME}:}{The ACS table's identification number}
\item{\code{• TOPIC}:}{The topic of the ACS table}
\item{\code{• UNIVERSE}:}{The universe that the ACS table's count represent
(e.g., households, adults, population living in occupied housing units, etc.)}

}}
\usage{
acs_tables
}
\description{
A \code{tibble} of the American Community Survey tables that
are used in the Neighborhood Change Typology model.
}
\seealso{
\code{\link{get_external_data_plan}}
}
\keyword{datasets}
