% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census-tracts-2016.R
\name{prepare_census_tracts_2016}
\alias{prepare_census_tracts_2016}
\alias{make_census_tracts_2016}
\title{Make King County's Census Tract Boundaries (2016)}
\usage{
prepare_census_tracts_2016(path)

make_census_tracts_2016(path)
}
\value{
a MULTIPOLYGON simple feature (class =  \code{sf})
}
\description{
Download the polygons for the census tracts in King County, WA
sing the \code{\link[tigris]{tigris}} package.
}
