% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-osf-projet-permission.R
\name{check_osf_access}
\alias{check_osf_access}
\title{Check Whether The User Has Access To An OSF Project}
\usage{
check_osf_access(project_title = NULL)
}
\arguments{
\item{project_title}{Character, the extact title of the OSF project}
}
\value{
a logical
}
\description{
Check whether the user has access to a specific OSF project.
}
