% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-to-2018-dollars.R
\name{convert_to_2018_dollars}
\alias{convert_to_2018_dollars}
\title{Convert to 2018-Dollars}
\usage{
convert_to_2018_dollars(value, date_string, cpi,
  series_title = "less_shelter")
}
\arguments{
\item{value}{numeric}

\item{date_string}{character, a date-like character string in the following pattern: "YYYY-MM-DD"}

\item{cpi}{list, the \code{cpi} list object}

\item{series_title}{character, either \code{"less_shelter"} (default) or \code{"all"}}
}
\value{
a named list (length == 2) contained named vectors of the inflation adjustment rate for each year.
}
\description{
Convert a dollar value to its inflation-adjusted 2018 dollar value.
Two time series from the US Bureau of Labor Statistics (BLS) are provided: all goods (CUSR0000SA0)
and all good less shelter (CUSR0000SA0L2).
}
\note{
A list of all the BLS series id codes is available here: \url{https://download.bls.gov/pub/time.series/cu/cu.series}
}
