% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi.R
\name{prepare_cpi_all}
\alias{prepare_cpi_all}
\alias{make_cpi_all}
\alias{prepare_cpi_less_shelter}
\alias{make_cpi_less_shelter}
\title{Make the CPI Key for Inflation-Adjusted Dollar Values}
\usage{
prepare_cpi_all(path)

make_cpi_all(path)

prepare_cpi_less_shelter(path)

make_cpi_less_shelter(path)
}
\value{
a named vector (\code{class = double})
}
\description{
Make the Consumer Price Index key for calculating inflation-adjusted dollar values.
}
