% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-geographies.R
\name{export_kc_boundary_gpkg}
\alias{export_kc_boundary_gpkg}
\alias{export_white_center_place_gpkg}
\alias{export_coo_communities_gpkg}
\title{Export Geographies As Files}
\usage{
export_kc_boundary_gpkg(kc_boundary, path)

export_white_center_place_gpkg(white_center_place, path)

export_coo_communities_gpkg(model_all, path)
}
\arguments{
\item{kc_boundary}{desc}

\item{path}{desc}

\item{white_center_place}{desc}

\item{model_all}{desc}
}
\description{
Thes functions write several spatial objects.
}
