% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exteral_datasets_list}
\alias{exteral_datasets_list}
\title{External Data List}
\format{The data itself is simply a \code{NULL} value.

Its purpose is to provide some documentation of the datasets that are not directly
included in the package.

The following external datasets are accessed by workflows in this package:

\describe{

\item{\code{• waterbodies}:}{Multipolygons of the major waterbodies in King County, WA (\href{https://osf.io/9fp6q/}{osf.io/9fp6q})}
\item{\code{• census_tracts_2009}:}{King County census tract boundaries (2009) (\href{https://osf.io/5gx8a/}{osf.io/5gx8a}).\cr
Original source: \href{ftp://ftp.census.gov/geo/tiger/TIGER2009/53_WASHINGTON/53033_King_County/tl_2009_53033_tract00.zip}{ftp.census.gov} }
\item{\code{• census_tracts_2016}:}{King County census tract boundaries (2009) \cr
Downloaded with the \code{\link[tigris]{tigris}} package. }
\item{\code{• acs_data}:}{A \code{tibble} of all of the American Community Survey data variables used in this project.\cr
Downloaded with the \code{\link[tidycensus]{tidycensus}} package.}
\item{\code{• parcel_value}:}{Tax value history for all parcels in King County, WA (\href{https://osf.io/XXX/}{osf.io/XXX})}

}}
\usage{
exteral_datasets_list
}
\description{
This is a list of datasets that are included in the package workflow but
are not stored in the \code{data/} directory.

The primary reason that a dataset is included in this list is that it exceeds Github's
50 MB file size limit (see \href{https://help.github.com/articles/conditions-for-large-files/}{here} for details).
}
\seealso{
\code{\link{get_external_data_plan}}
}
\keyword{datasets}
