% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factfinder-data.R
\name{prepare_factfinder_data}
\alias{prepare_factfinder_data}
\alias{make_factfinder_data}
\title{Make A Tibble of The Project's US Census Variables from American Factfinder}
\usage{
prepare_factfinder_data(indicator_template, acs_tables, path)

make_factfinder_data(path)
}
\arguments{
\item{indicator_template}{Tibble, the \code{indicator_template} object}

\item{acs_tables}{Tibble, the \code{acs_table} object}

\item{path}{Character, the path or connection to write to.}
}
\value{
a \code{tibble}
}
\description{
Return a \code{tibble} of all of the US Census data variables
that are obtained from the American Factfinder user interface: \link{https://factfinder.census.gov/}
}
\note{
Data source: \link{https://factfinder.census.gov/faces/nav/jsf/pages/searchresults.xhtml?refresh=t}
}
