% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-bls-data.R
\name{get_bls_data}
\alias{get_bls_data}
\title{Get BLS Data}
\usage{
get_bls_data(..., bls_api_key = "")
}
\arguments{
\item{...}{Arguments to be passed on to \code{blsAPI::\link[blsAPI]{blsAPI}}}

\item{bls_api_key}{Character, the API key provided to you from the Census formated in quotes}
}
\value{
a JSON string or a data.frame
}
\description{
Allows users to request data for one or multiple series through the U.S. Bureau of Labor Statistics API.
Users provide parameters as specified in \url{https://www.bls.gov/developers/api_signature.htm}
and the function returns a JSON string or data frame.
}
\seealso{
\code{\link[blsAPI]{blsAPI}}
}
