% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{get_data_source_plan}
\alias{get_data_source_plan}
\title{Get the Data Plans}
\usage{
get_data_source_plan()
}
\value{
a \code{drake} plan
}
\description{
Use \code{\link[drake]{drake_plan}} to create the data plans.
}
\examples{

# Print one of the external data plans

get_data_cache_plan()


# Make the plan, load a target, print the target

\dontrun{

data_plans <- bind_plans(get_data_source_plan(), get_data_cache_plan())

make(data_plans)

readd(kc_boundary)
}
}
