% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{get_export_plan}
\alias{get_export_plan}
\title{Get the Export Plan}
\usage{
get_export_plan()
}
\value{
a \code{drake} plan
}
\description{
Use \code{\link[drake]{drake_plan}} to create the project's
templates plan.
}
