% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{get_external_data_plan}
\alias{get_external_data_plan}
\title{Get the External Data Plan}
\usage{
get_external_data_plan()
}
\value{
a \code{drake} plan
}
\description{
Use \code{\link[drake]{drake_plan}} to create the external data plan.
}
\examples{

# Print the plan

get_external_data_plan()


# Make the plan, load a target, print the target

\dontrun{

make(get_external_data_plan())

loadd(kc_boundary)

print(kc_boundary)
}
}
