% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\docType{data}
\name{get_indicator_plan}
\alias{get_indicator_plan}
\title{Get the Indicator Plan}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 2 columns.}
\usage{
get_indicator_plan
}
\value{
a \code{drake} plan
}
\description{
Use \code{\link[drake]{drake_plan}} to create the indicator plan.
}
\examples{

# Print the plan

get_indicator_plan()


# Make the plan, load a target, print the target

\dontrun{

make(get_indicator_plan())

loadd(kc_boundary)

print(kc_boundary)
}
}
\keyword{datasets}
