% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{get_model_plan}
\alias{get_model_plan}
\title{Get the Model Plan}
\usage{
get_model_plan()
}
\value{
a \code{drake} plan
}
\description{
Use \code{\link[drake]{drake_plan}} to create the typology plan.
}
\examples{

# Print the plan

get_model_plan()


# Make the plan, load a target, print the target

\dontrun{

make(get_model_plan())

loadd(acs_indicators)

print(acs_indicators)
}
}
