% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-osf-version.R
\name{get_osf_version}
\alias{get_osf_version}
\title{Get OSF File Version Number}
\usage{
get_osf_version(id = NULL, filename = NULL)
}
\arguments{
\item{id}{Character, OSF id (osf.io/XXXX) for the node (project or component)
to get the file info for}

\item{filename}{Character, the extact name of the specifc file}
}
\value{
a double
}
\description{
Get the version number of a file that is stored in a OSF project.
}
