% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{get_templates_plan}
\alias{get_templates_plan}
\title{Get the Project Templates Plan}
\usage{
get_templates_plan()
}
\value{
a \code{drake} plan
}
\description{
Use \code{\link[drake]{drake_plan}} to create the project's
templates plan.
}
\examples{

# Print one of the external data plans

get_templates_plan()


# Make the plan, load a target, print the target

\dontrun{

make(get_templates_plan())

readd(preliminary_model_plan)

}
}
