% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{get_variable_plan}
\alias{get_variable_plan}
\title{Get the Variable Plan}
\usage{
get_variable_plan()
}
\value{
a \code{drake} plan
}
\description{
Use \code{\link[drake]{drake_plan}} to create the variable plan.
}
\examples{

# Print the plan

get_variable_plan()


# Make the plan, load a target, print the target

\dontrun{

make(get_variable_plan())

loadd(parcel_tract_overlay)

print(parcel_tract_overlay)
}
}
