% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud-chas-data.R, R/ltdb-data.R
\name{prepare_hud_chas_data}
\alias{prepare_hud_chas_data}
\alias{make_hud_chas_data}
\alias{make_ltdb_data}
\title{Make A Tibble of The Project's HUD CHAS Variables}
\usage{
prepare_hud_chas_data(path)

make_hud_chas_data(path)

make_ltdb_data(path)
}
\arguments{
\item{path}{Character, the path or connection to write to.}
}
\value{
a \code{tibble}
}
\description{
Return a \code{tibble} of all of the HUD CHAS data variables
that are used in the Neighborhood Change Typology model for both 5-year spans
(2006-2010 and 2011-2015).
}
\note{
Data source: \link{https://www.huduser.gov/portal/datasets/cp.html}
}
