% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud-chas-variables.R
\name{make_hud_chas_variables}
\alias{make_hud_chas_variables}
\title{Make A Tibble of The Project's HUD CHAS Variables}
\usage{
make_hud_chas_variables(hud_chas_data, hud_chas_data_lut, model_table,
  variable_template)
}
\arguments{
\item{hud_chas_data}{Tibble, desc}

\item{hud_chas_data_lut}{Tibble, desc}

\item{model_table}{Tibble, desc}

\item{variable_template}{Tibble, desc}
}
\value{
a \code{tibble}
}
\description{
Return a \code{tibble} of all of the HUD CHAS data variables
that are used in the Neighborhood Change Typology model for both 5-year spans
(2006-2010 and 2011-2015).
}
