% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc-boundary.R
\name{prepare_kc_boundary}
\alias{prepare_kc_boundary}
\alias{make_kc_boundary}
\title{Make the Boundary of King County, WA}
\usage{
prepare_kc_boundary(path)

make_kc_boundary(path)
}
\value{
a MULTIPOLYGON simple feature (class =  \code{sf})
}
\description{
Get the polygon of King County, WA using the \code{\link[tigris]{tigris}} package.
}
