% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kc_boundary}
\alias{kc_boundary}
\title{The Boundary of King County, WA}
\format{The data is a single record simple feature (class = \code{sf}) with
one variable and a \code{MULTIPOLYGON} geometry column.

The object includes the following variables:
\describe{

\item{\code{• NAME}:}{The name of the county}
\item{\code{• geometry}:}{The simple feature column (geometry type is \code{MULTIPOLYGON})}

}}
\usage{
kc_boundary
}
\description{
A polygon of the boundary of King County, WA.
}
\seealso{
\code{\link{get_external_data_plan}}
}
\keyword{datasets}
