% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltdb-data.R
\name{prepare_ltdb_data}
\alias{prepare_ltdb_data}
\alias{make_ltdb_data}
\title{Make A Tibble of The Project's US Census Variables from LTDB}
\usage{
prepare_ltdb_data(indicator_template, acs_tables, path)

make_ltdb_data(path)
}
\arguments{
\item{indicator_template}{Tibble, the \code{indicator_template} object}

\item{acs_tables}{Tibble, the \code{acs_table} object}

\item{path}{Character, the path or connection to write to.}
}
\value{
a \code{tibble}
}
\description{
Return a \code{tibble} of all of the US Census data variables
that are obtained from the Brown University Longitudinal Tract Database (LTDB).
}
\note{
Data source: \link{https://s4.ad.brown.edu/projects/diversity/Researcher/LTBDDload/DataList.aspx}
}
