% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-acs-data.R
\name{make_acs_data}
\alias{make_acs_data}
\title{Make A Tibble of The Project's ACS Variables}
\usage{
make_acs_data(acs_tables)
}
\arguments{
\item{acs_tables}{Tibble, the \code{acs_table} object}
}
\value{
a \code{tibble}
}
\description{
Return a \code{tibble} of all of the American Community Survey data variables
that are used in the Neighborhood Change Typology model for both 5-year spans
(2006-2010 and 2012-2016).
}
\seealso{
make_acs_table
}
