% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-acs-indicators.R
\name{make_acs_indicators}
\alias{make_acs_indicators}
\title{Make The Census Data (ACS) Indicators}
\usage{
make_acs_indicators(acs_data, acs_tables)
}
\arguments{
\item{acs_data}{desc}

\item{acs_tables}{desc}
}
\value{
a \code{tibble}
}
\description{
Make The Census Data (ACS) Indicators
}
