% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-acs-tables.R
\name{make_acs_tables}
\alias{make_acs_tables}
\title{Make A Tibble of ACS Tables}
\usage{
make_acs_tables()
}
\value{
a \code{tibble}
}
\description{
Return a \code{tibble} of the American Community Survey tables that
are used in the Neighborhood Change Typology model.
}
