% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-census-tract-neighbors.R
\name{make_census_tract_neighbors}
\alias{make_census_tract_neighbors}
\title{Make the Census Tract Neighbors}
\usage{
make_census_tract_neighbors(census_tracts_2016)
}
\arguments{
\item{census_tracts_2016}{Tibble, Temporary description.}
}
\value{
a simple feature object (class = \code{sf})
}
\description{
Temporary description.
}
