% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-census-tracts-2009.R
\name{make_census_tracts_2009}
\alias{make_census_tracts_2009}
\title{Make King County's Census Tract Boundaries (2009)}
\usage{
make_census_tracts_2009()
}
\value{
a MULTIPOLYGON simple feature (class =  \code{sf})
}
\description{
Download the polygons for the census tracts in King County, WA from osf.io.
}
