% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-change-dategroupid-long.R
\name{make_change_dategroupid_long}
\alias{make_change_dategroupid_long}
\title{Make The Change Date Group IDs (Long)}
\usage{
make_change_dategroupid_long(model_table_production)
}
\arguments{
\item{model_table_production}{desc}
}
\value{
a \code{tibble}
}
\description{
Description
}
