% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-condo-criteria.R
\name{make_condo_criteria}
\alias{make_condo_criteria}
\title{Make the Criteria for Condominium Units}
\usage{
make_condo_criteria(condo_unit_type_key)
}
\value{
a list
}
\description{
Make the criteria used to identify the condominium
units that will be used to identify housing market characteristics.
}
