% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condo-info-2005.R
\name{make_condo_info_2005}
\alias{make_condo_info_2005}
\title{Make The General Tax Data For King County Condominium Units (2005)}
\usage{
make_condo_info_2005(zip_path, file_path)
}
\arguments{
\item{zip_path}{Character, the file path of the archive file}

\item{file_path}{Character, the file path of the file that is to be extracted}
}
\value{
a \code{tibble}
}
\description{
Make the general tax data for condominium units in King County, Washington.
}
