% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-condo-unit-type-key.R
\name{make_condo_unit_type_key}
\alias{make_condo_unit_type_key}
\title{Make the Key For Decoding Condo Unit Categories}
\usage{
make_condo_unit_type_key(parcel_lut_2005, parcel_lut_2018)
}
\arguments{
\item{parcel_lut_2005}{Tibble, The lookup table data for the 2005
King County Assessor parcel data.}

\item{parcel_lut_2018}{Tibble, The lookup table data for the 2018
King County Assessor parcel data.}
}
\value{
a \code{tibble}
}
\description{
Make the key for decoding the King County Assessor's
"Condo Unit" categories.
}
