% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-cpi.R
\name{make_cpi}
\alias{make_cpi}
\title{Make the CPI Key for Inflation-Adjusted Dollar Values}
\usage{
make_cpi()
}
\value{
a named vector (\code{class = double})
}
\description{
Make the Consumer Price Index key for calculating inflation-adjusted dollar values.
}
