% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-dl-parcel-data.R
\name{make_dl_parcel_data}
\alias{make_dl_parcel_data}
\title{Download the Zipped King County Parcels Data}
\usage{
make_dl_parcel_data()
}
\value{
\code{NULL}
}
\description{
Download a compressed archive of the tax assessor's data for
all parcels in King County, WA (1998 - 2017).
}
