% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-excluded-tract-geoids.R
\name{make_excluded_tract_geoids}
\alias{make_excluded_tract_geoids}
\title{Make The List of Excluded Census Tracts}
\usage{
make_excluded_tract_geoids()
}
\value{
a character string
}
\description{
Temporary description
}
