% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-housing-market-indicators.R
\name{make_housing_market_indicators}
\alias{make_housing_market_indicators}
\title{Make Housing Market Indicators}
\usage{
make_housing_market_indicators(census_tracts_2016, excluded_tract_geoids,
  parcel_boundaries, parcel_tract_overlay, housing_market_parcel_value,
  housing_market_parcel_appr)
}
\arguments{
\item{census_tracts_2016}{Simple feature, Temporary description}

\item{excluded_tract_geoids}{Character string, Temporary description}

\item{parcel_boundaries}{Simple feature, Temporary description}

\item{parcel_tract_overlay}{Tibble, Temporary description}

\item{housing_market_parcel_value}{Tibble, Temporary description}

\item{housing_market_parcel_appr}{Tibble, Temporary description}
}
\value{
a simple feature
}
\description{
Temporary description
}
