% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-housing-market-parcel-appreciation.R
\name{make_housing_market_parcel_appr}
\alias{make_housing_market_parcel_appr}
\title{Make Housing Market Parcel Value Appreciation Data}
\usage{
make_housing_market_parcel_appr(housing_market_parcel_value)
}
\arguments{
\item{present_use_key}{Tibble, Temporary description}

\item{single_family_criteria}{Tibble, Temporary description.}

\item{cpi}{Tibble, Temporary description.}

\item{parcel_value}{Tibble, Temporary description.}

\item{parcel_info_2005}{Tibble, Temporary description.}

\item{parcel_info_2010}{Tibble, Temporary description.}

\item{parcel_info_2018}{Tibble, Temporary description.}
}
\value{
a \code{tibble}
}
\description{
Temporary description
}
