% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-housing-market-parcel-value.R
\name{make_housing_market_parcel_value}
\alias{make_housing_market_parcel_value}
\title{Make Housing Market Parcel Value Data}
\usage{
make_housing_market_parcel_value(present_use_key, condo_unit_type_key,
  single_family_criteria, condo_criteria, cpi, parcel_value, parcel_info_2005,
  parcel_info_2010, parcel_info_2018, condo_info_2005, condo_info_2010,
  condo_info_2018)
}
\arguments{
\item{present_use_key}{Tibble, Temporary description}

\item{condo_unit_type_key}{Tibble, Temporary description}

\item{single_family_criteria}{Tibble, Temporary description.}

\item{condo_criteria}{Tibble, Temporary description.}

\item{cpi}{Tibble, Temporary description.}

\item{parcel_value}{Tibble, Temporary description.}

\item{parcel_info_2005}{Tibble, Temporary description.}

\item{parcel_info_2010}{Tibble, Temporary description.}

\item{parcel_info_2018}{Tibble, Temporary description.}

\item{condo_info_2005}{Tibble, Temporary description.}

\item{condo_info_2010}{Tibble, Temporary description.}

\item{condo_info_2018}{Tibble, Temporary description.}
}
\value{
a \code{tibble}
}
\description{
Temporary description
}
