% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-housing-market-parcels.R
\name{make_housing_market_parcels}
\alias{make_housing_market_parcels}
\title{Make Housing Market Parcel Data}
\usage{
make_housing_market_parcels(present_use_key, single_family_criteria, cpi,
  parcel_value, parcel_info_2005, parcel_info_2010, parcel_info_2018)
}
\arguments{
\item{present_use_key}{Tibble, Temporary description}

\item{single_family_criteria}{Tibble, Temporary description.}

\item{cpi}{Tibble, Temporary description.}

\item{parcel_value}{Tibble, Temporary description.}

\item{parcel_info_2005}{Tibble, Temporary description.}

\item{parcel_info_2010}{Tibble, Temporary description.}

\item{parcel_info_2018}{Tibble, Temporary description.}
}
\value{
a \code{tibble}
}
\description{
Temporary description
}
