% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-housing-market-sales.R
\name{make_housing_market_sales}
\alias{make_housing_market_sales}
\title{Make Housing Market Sales Velocity Data}
\usage{
make_housing_market_sales(parcel_sales, sales_lut_key_list, sales_criteria,
  present_use_key, single_family_criteria, condo_unit_type_key, condo_criteria,
  cpi, parcel_info_2005, parcel_info_2010, parcel_info_2018, condo_info_2005,
  condo_info_2010, condo_info_2018, res_bldg_2005, res_bldg_2010, res_bldg_2018)
}
\arguments{
\item{parcel_sales}{Tibble, Temporary description}

\item{sales_lut_key_list}{Tibble, Temporary description}

\item{sales_criteria}{Tibble, Temporary description}

\item{present_use_key}{Tibble, Temporary description}

\item{single_family_criteria}{Tibble, Temporary description}

\item{condo_unit_type_key}{Tibble, Temporary description}

\item{condo_criteria}{Tibble, Temporary description}

\item{cpi}{Tibble, Temporary description}

\item{parcel_info_2005}{Tibble, Temporary description.}

\item{parcel_info_2010}{Tibble, Temporary description.}

\item{parcel_info_2018}{Tibble, Temporary description.}

\item{condo_info_2005}{Tibble, Temporary description.}

\item{condo_info_2010}{Tibble, Temporary description.}

\item{condo_info_2018}{Tibble, Temporary description.}

\item{res_bldg_2005}{Tibble, Temporary description}

\item{res_bldg_2010}{Tibble, Temporary description}

\item{res_bldg_2018}{Tibble, Temporary description}
}
\value{
a \code{tibble}
}
\description{
Temporary description
}
