% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-housing-market-sale-velocity.R
\name{make_housing_market_sales_velocity}
\alias{make_housing_market_sales_velocity}
\title{Make Housing Market Sales Velocity Data}
\usage{
make_housing_market_sales_velocity(housing_market_sales)
}
\arguments{
\item{housing_market_sales}{Tibble, Temporary description}
}
\value{
a \code{tibble}
}
\description{
Temporary description
}
