% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-indicator-template.R
\name{make_indicator_template}
\alias{make_indicator_template}
\title{Make A Template for Indicator Objects}
\usage{
make_indicator_template()
}
\value{
a \code{tibble}
}
\description{
Return a \code{tibble} containing the unified set of columns and
column types used for all indicator objects in this project.
}
