% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-indicators.R
\name{make_indicators}
\alias{make_indicators}
\title{Make All Indicators}
\usage{
make_indicators(indicators_cnt_pct, indicators_median, sample_size_metadata)
}
\arguments{
\item{indicators_cnt_pct}{desc}

\item{indicators_median}{desc}

\item{sample_size_metadata}{desc}
}
\value{
a \code{tibble}
}
\description{
Desc
}
