% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-indicators-by-dimension.R
\name{make_indicators_by_DIMENSION}
\alias{make_indicators_by_DIMENSION}
\title{Make The Indicators by DIMENSION}
\usage{
make_indicators_by_DIMENSION(indicators_cnt_pct, indicators_median,
  model_table_inputs, indicator_DIMENSION_template)
}
\arguments{
\item{indicators_cnt_pct}{desc}

\item{indicators_median}{desc}

\item{model_table_inputs}{desc}

\item{indicator_DIMENSION_template}{desc}
}
\value{
a \code{tibble}
}
\description{
Description
}
