% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-indicators-cnt-pct.R
\name{make_indicators_cnt_pct}
\alias{make_indicators_cnt_pct}
\title{Make The Count Indicators}
\usage{
make_indicators_cnt_pct(acs_variables, hud_chas_variables)
}
\arguments{
\item{acs_variables}{desc}

\item{hud_chas_variables}{desc}
}
\value{
a \code{tibble}
}
\description{
Make the ACS Inidcators related to a \emph{count} of a population.
An example of the type of indicator included in this object might be
the count of renter households, while the median rent price would \emph{not} be included.
}
