% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-indicators-in-models.R
\name{make_indicators_in_models}
\alias{make_indicators_in_models}
\title{Make The Indicators in Models}
\usage{
make_indicators_in_models(indicators_by_dimension, model_table_production)
}
\arguments{
\item{indicators_by_dimension}{desc}

\item{model_table_production}{desc}
}
\value{
a \code{tibble}
}
\description{
Description
}
