% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-indicators-pct.R
\name{make_indicators_pct}
\alias{make_indicators_pct}
\title{Make The Census Data (ACS) Indicators}
\usage{
make_indicators_pct(acs_data, hud_chas_data, acs_tables)
}
\arguments{
\item{acs_data}{desc}

\item{hud_chas_data}{desc}

\item{acs_tables}{desc}
}
\value{
a \code{tibble}
}
\description{
Make the ACS Inidcators related to a \emph{proportion} of a population.
An example of the type of indicator included in this object might be
the percentage of renter households, while the median rent price would \emph{not} be included.
}
