% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-indicators-proximity.R
\name{make_indicators_proximity}
\alias{make_indicators_proximity}
\title{Make The Proximity Indicators}
\usage{
make_indicators_proximity(census_tracts_2016_trimmed, indicator_type_template)
}
\arguments{
\item{census_tracts_2016_trimmed}{desc}

\item{indicator_type_template}{desc}
}
\value{
a \code{tibble}
}
\description{
Description
}
