% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-indicators-values.R
\name{make_indicators_values}
\alias{make_indicators_values}
\title{Make The Value Indicators}
\usage{
make_indicators_values(acs_variables, ltdb_variables, factfinder_variables,
  parcel_value_variables, parcel_sales_variables, parcel_tract_overlay,
  indicator_template)
}
\arguments{
\item{acs_variables}{desc}

\item{ltdb_variables}{desc}

\item{factfinder_variables}{desc}

\item{parcel_value_variables}{desc}

\item{parcel_sales_variables}{desc}

\item{parcel_tract_overlay}{desc}
}
\value{
a \code{tibble}
}
\description{
Make the ACS Inidcators related to a _value of a population.
An example of the type of indicator included in this object might be
the count of renter households, while the median rent price would \emph{not} be included.
}
