% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-indicators-wide.R
\name{make_indicators_wide}
\alias{make_indicators_wide}
\title{Make The Project Inidicators (Wide)}
\usage{
make_indicators_wide(model_table_column_type, model_table_production,
  community_metadata, indicators_comparison, indicators_comparison_of_change,
  indicators_change_in_comparison, indicators_proximity)
}
\arguments{
\item{model_table_column_type}{desc}

\item{model_table_production}{desc}

\item{community_metadata}{desc}

\item{indicators_comparison}{desc}

\item{indicators_comparison_of_change}{desc}

\item{indicators_change_in_comparison}{desc}

\item{indicators_proximity}{desc}
}
\value{
a \code{tibble}
}
\description{
Description
}
