% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-kc-boundary.R
\name{make_kc_boundary}
\alias{make_kc_boundary}
\title{Make the Boundary of King County, WA}
\usage{
make_kc_boundary()
}
\value{
a MULTIPOLYGON simple feature (class =  \code{sf})
}
\description{
Get the polygon of King County, WA using the \code{\link[tigris]{tigris}} package.
}
