% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-models.R
\name{make_model_vulnerability}
\alias{make_model_vulnerability}
\title{Make The Models}
\usage{
make_model_vulnerability(model_table, indicators_comparison)
}
\arguments{
\item{model_table}{desc}

\item{indicators_comparison}{Tibble, Temporary description.}

\item{indicators_comparison_of_change}{desc}

\item{indicators_change_in_comparison}{desc}
}
\value{
a \code{tibble}
}
\description{
Temporary description.
}
