% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-king-county-parcel-boundaries.R
\name{make_parcel_boundaries}
\alias{make_parcel_boundaries}
\title{Make King County's Parcel Boundaries (2018)}
\usage{
make_parcel_boundaries()
}
\value{
a MULTIPOLYGON simple feature (class =  \code{sf})
}
\description{
Download the polygons for the tax parcels in King County, WA from osf.io.
}
