% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel-info-2005.R
\name{make_parcel_info_2005}
\alias{make_parcel_info_2005}
\title{Make The General Tax Data For King County Parcels (2005)}
\usage{
make_parcel_info_2005(zip_path, file_path)
}
\arguments{
\item{zip_path}{Character, the file path of the archive file}

\item{file_path}{Character, the file path of the file that is to be extracted}
}
\value{
a \code{tibble}
}
\description{
Make the general tax data for King County parcels
(Tax Year 2005, Calendar Year 2004)
}
