% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel-info-2010.R
\name{make_parcel_info_2010}
\alias{make_parcel_info_2010}
\title{Make The General Tax Data For King County Parcels (2010)}
\usage{
make_parcel_info_2010(zip_path, file_path)
}
\arguments{
\item{zip_path}{Character, the file path of the archive file}

\item{file_path}{Character, the file path of the file that is to be extracted}
}
\value{
a \code{tibble}
}
\description{
Make the general tax data for King County parcels
(Tax Year 2010, Calendar Year 2009)
}
