% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-parcels-tract-overlay.R
\name{make_parcel_tract_overlay}
\alias{make_parcel_tract_overlay}
\title{Make Census Tract to Parcel Overlay}
\usage{
make_parcel_tract_overlay(parcel_boundaries, census_tracts_2016)
}
\arguments{
\item{parcel_boundaries}{Simple feature, Temporary description}

\item{census_tracts_2016}{Simple feature, Temporary description.}
}
\value{
a simple feature (class = \code{sf})
}
\description{
Temporary description
}
