% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-pin.R
\name{make_pin}
\alias{make_pin}
\title{Make a Parcel Identification Number}
\usage{
make_pin(major, minor)
}
\arguments{
\item{major}{Character, the first six (6) digits of King County's
ten-digit Parcel Identification number}

\item{minor}{Character, the last four (4) digits of King County's
ten-digit Parcel Identification number}
}
\value{
a \code{tibble}
}
\description{
Make a complete ten-digit Parcel Identification Number.
}
