% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-model-portland.R
\name{make_portland_model_vulnerability}
\alias{make_portland_model_vulnerability}
\title{Make The Portland Model}
\usage{
make_portland_model_vulnerability(indicators_comparison)
}
\arguments{
\item{indicators_comparison}{Tibble, Temporary description.}
}
\value{
a \code{tibble}
}
\description{
Temporary description.
}
