% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-present-use-key.R
\name{make_present_use_key}
\alias{make_present_use_key}
\title{Make the Key For Decoding Present Use Categories}
\usage{
make_present_use_key(parcel_lut_2005, parcel_lut_2018)
}
\arguments{
\item{parcel_lut_2005}{Tibble, The lookup table data for the 2005
King County Assessor parcel data.}

\item{parcel_lut_2018}{Tibble, The lookup table data for the 2018
King County Assessor parcel data.}
}
\value{
a \code{tibble}
}
\description{
Make the key for decoding the King County Assessor's
"Present Use" categories.
}
