% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-previous-typology.R
\name{make_previous_typology}
\alias{make_previous_typology}
\title{Make Previous Typology}
\usage{
make_previous_typology()
}
\value{
a MULTIPOLYGON simple feature (class =  \code{sf})
}
\description{
This is a temporary function.
}
