% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-sales-criteria.R
\name{make_sales_criteria}
\alias{make_sales_criteria}
\title{Make the Criteria for Real Property Sales}
\usage{
make_sales_criteria()
}
\value{
a named list
}
\description{
Make the criteria used to identify the real propert sales
that will be used to identify housing market characteristics.
}
