% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-sales-lut-key-list.R
\name{make_sales_lut_key_list}
\alias{make_sales_lut_key_list}
\title{Make Housing Market Sales Lookup Table Key List}
\usage{
make_sales_lut_key_list(parcel_lut_2018)
}
\arguments{
\item{parcel_lut_2018}{Temporary description}
}
\value{
A named list of tibbles
}
\description{
Temporary description
}
