% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-sample-size-metadata.R
\name{make_sample_size_metadata}
\alias{make_sample_size_metadata}
\title{Make The Indicator Sample Size Metadata}
\usage{
make_sample_size_metadata(indicators_cnt_pct, indicators_median)
}
\arguments{
\item{indicators_cnt_pct}{desc}

\item{indicators_median}{desc}
}
\value{
a \code{tibble}
}
\description{
Description
}
