% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-single-family-criteria.R
\name{make_single_family_criteria}
\alias{make_single_family_criteria}
\title{Make the Criteria for Single Family Housing Parcels}
\usage{
make_single_family_criteria(present_use_key)
}
\value{
a list
}
\description{
Make the criteria used to identify the single family
parcels that will be used to identify housing market characteristics.
}
