% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-typology.R
\name{make_typology}
\alias{make_typology}
\title{Make  Typology}
\usage{
make_typology(vulnerability_indicators, demo_change_indicators,
  previous_typology)
}
\value{
a MULTIPOLYGON simple feature (class =  \code{sf})
}
\description{
This is a temporary function.
}
