% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-waterbodies.R
\name{make_waterbodies}
\alias{make_waterbodies}
\title{Make King County's Major Waterbodies}
\usage{
make_waterbodies(kc_boundary)
}
\arguments{
\item{kc_boundary}{Polygon, the boundary of King County, WA}
}
\value{
a MULTIPOLYGON simple feature (class =  \code{sf})
}
\description{
Download the polygons for the major waterbodies in Washington State from osf.io
and filter the data to include those waterbodies within King County.
}
