% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-white-center-place.R
\name{make_white_center_place}
\alias{make_white_center_place}
\title{Make White Center Place}
\usage{
make_white_center_place()
}
\value{
a MULTIPOLYGON simple feature (class =  \code{sf})
}
\description{
This is a temporary function.
}
