% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-variables.R
\name{make_parcel_all_metadata}
\alias{make_parcel_all_metadata}
\alias{make_county_tract_all_metadata}
\title{Prepare Variable Metadata Objects}
\usage{
make_parcel_all_metadata(present_use_key, condo_unit_type_key,
  parcel_tract_overlay, parcel_info_2005, parcel_info_2010, parcel_info_2018,
  condo_info_2005, condo_info_2010, condo_info_2018, res_bldg_2005,
  res_bldg_2010, res_bldg_2018, variable_template)

make_county_tract_all_metadata(acs_data)
}
\arguments{
\item{present_use_key}{desc}

\item{condo_unit_type_key}{desc}

\item{parcel_tract_overlay}{desc}

\item{parcel_info_2005}{desc}

\item{parcel_info_2010}{desc}

\item{parcel_info_2018}{desc}

\item{condo_info_2005}{desc}

\item{condo_info_2010}{desc}

\item{condo_info_2018}{desc}

\item{res_bldg_2005}{desc}

\item{res_bldg_2010}{desc}

\item{res_bldg_2018}{desc}

\item{variable_template}{desc}

\item{acs_data}{desc}

\item{res_bldg_2018}{desc}
}
\value{
a \code{tibble}
}
\description{
Return a \code{tibble} object containing the metadata used for all o.
}
