% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-model-table.R
\name{make_model_table_inputs}
\alias{make_model_table_inputs}
\alias{make_model_table}
\title{Make A Tibble of the Project Models}
\usage{
make_model_table_inputs()

make_model_table()
}
\value{
a \code{tibble}
}
\description{
Return a \code{tibble} containing metadata for each of the
models used to describe neighborhood change.
}
