% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-models.R
\name{make_model_pdx16}
\alias{make_model_pdx16}
\title{Make The Models}
\usage{
make_model_pdx16(indicators_wide, census_tracts_2016_trimmed)
}
\arguments{
\item{indicators_wide}{desc}
}
\value{
a \code{tibble}
}
\description{
Temporary description.
}
