% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf-download-files.R
\name{osf_download_files}
\alias{osf_download_files}
\title{Download Files From OSF}
\usage{
osf_download_files(id, path = NULL, private = FALSE, version = NULL)
}
\arguments{
\item{id}{Character, OSF id (osf.io/xxxx; just xxxx).}

\item{path}{Character, Path to file on local machine to upload.}

\item{private}{Logcial, Whether the file is private.}

\item{version}{Character, The OSF version id.}
}
\value{
the filepath of the downloaded file
}
\description{
Download files from the OSF.
}
