% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf-upload-or-update.R
\name{osf_upload_or_update}
\alias{osf_upload_or_update}
\title{Upload/Update A File On OSF}
\usage{
osf_upload_or_update(has_osf_access = FALSE, project_id = NULL,
  file_id = NULL, path, dest = NULL)
}
\arguments{
\item{has_osf_access}{Logical, Whether the user has access to the project's OSF repository.}

\item{project_id}{Character, The id of the OSF project.}

\item{file_id}{Character, The id of the existing OSF file (if any).}

\item{path}{Character, The file path.}

\item{dest}{Character, The name of the destination file on OSF (if NULL, basename(path) will be used).}
}
\value{
a logical
}
