% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel-boundaries.R
\name{prepare_parcel_boundaries}
\alias{prepare_parcel_boundaries}
\alias{make_parcel_boundaries}
\title{Make King County's Parcel Boundaries (2018)}
\usage{
prepare_parcel_boundaries(path)

make_parcel_boundaries(path)
}
\value{
a MULTIPOLYGON simple feature (class =  \code{sf})
}
\description{
Download the polygons for the tax parcels in King County, WA from osf.io.
}
