% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\docType{data}
\name{plan_external_data}
\alias{plan_external_data}
\title{Drake Plan: External Data}
\format{A \code{\link[drake]{drake_plan}} object of class \code{tbl_df} that builds
the following \code{\link[drake]{target}}s:
\describe{

\item{\code{• kc_boundary}:}{A polygon of King County, WA}

}}
\usage{
plan_external_data
}
\description{
A \code{\link[drake]{drake_plan}} that loads several external datasets.
}
\examples{

# Print the plan

print(plan_external_data)


# Make the plan, load a target, print the target

\dontrun{

make(plan_external_data)

readd(kc_boundary)
}
}
\seealso{
\code{\link{exteral_datasets_list}}
}
\keyword{drake,}
\keyword{plan}
