% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-project-tables.R
\name{make_acs_tables}
\alias{make_acs_tables}
\alias{make_model_table_inputs}
\alias{make_model_table_production}
\title{Make A Tibble of the Project Models}
\usage{
make_acs_tables()

make_model_table_inputs(path)

make_model_table_production(path)
}
\value{
a \code{tibble}
}
\description{
Return a \code{tibble} containing metadata for each of the
models used to describe neighborhood change.
}
