% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-typology-map.R
\name{show_typology_map}
\alias{show_typology_map}
\alias{show_pdx18_sf_map}
\alias{show_coo16_sf_map}
\alias{show_coo18_sf_map}
\alias{show_coorev18_mf_map}
\alias{show_coorev18_sf_map}
\title{SHow Typology Map}
\usage{
show_typology_map()

show_pdx18_sf_map()

show_coo16_sf_map()

show_coo18_sf_map()

show_coorev18_mf_map()

show_coorev18_sf_map()
}
\arguments{
\item{model_all}{desc}

\item{kc_boundary}{desc}
}
\value{
a leaflet map
}
\description{
This is a temporary function.
}
