% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-typology-map.R
\name{show_typology_map}
\alias{show_typology_map}
\title{SHow Typology Map}
\usage{
show_typology_map(typology, white_center_place)
}
\arguments{
\item{typology}{typology}

\item{white_center_place}{white center boundary}
}
\value{
a MULTIPOLYGON simple feature (class =  \code{sf})
}
\description{
This is a temporary function.
}
