% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload-revision.R
\name{upload_revision}
\alias{upload_revision}
\title{Upload Revision to OSF}
\usage{
upload_revision(id, path, ...)
}
\arguments{
\item{id}{Character, OSF id (osf.io/xxxx; just xxxx).}

\item{path}{Character, Path to file on local machine to upload.}

\item{...}{Other arguments.}
}
\value{
the url of the revised file
}
\description{
Upload Revision to OSF
}
