% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white-center-place.R
\name{prepare_white_center_place}
\alias{prepare_white_center_place}
\alias{make_white_center_place}
\title{Make White Center Place}
\usage{
prepare_white_center_place(path)

make_white_center_place(path)
}
\value{
a MULTIPOLYGON simple feature (class =  \code{sf})
}
\description{
This is a temporary function.
}
