% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-data.R
\name{prepare_acs_data}
\alias{prepare_acs_data}
\alias{make_acs_data}
\title{Make A Tibble of The Project's ACS Variables}
\usage{
prepare_acs_data(model_table, acs_tables, path)

make_acs_data(path)
}
\arguments{
\item{model_table}{Tibble, the \code{model_table} object}

\item{acs_tables}{Tibble, the \code{acs_table} object}

\item{path}{Character, the path or connection to write to.}
}
\value{
a \code{tibble}
}
\description{
Return a \code{tibble} of all of the American Community Survey data variables
that are used in the Neighborhood Change Typology model for both 5-year spans
(2006-2010 and 2012-2016).
}
\seealso{
make_acs_table
}
