% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi.R
\name{prepare_cpi}
\alias{prepare_cpi}
\alias{make_cpi}
\title{Make the CPI Key for Inflation-Adjusted Dollar Values}
\usage{
prepare_cpi(path)

make_cpi(path)
}
\value{
a named vector (\code{class = double})
}
\description{
Make the Consumer Price Index key for calculating inflation-adjusted dollar values.
}
