% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-model.R
\name{export_model_all_csv}
\alias{export_model_all_csv}
\alias{export_model_all_gpkg}
\alias{export_model_typologies_csv}
\alias{export_model_typologies_gpkg}
\title{Export Model As Files}
\usage{
export_model_all_csv(model_all, path)

export_model_all_gpkg(model_all, path)

export_model_typologies_csv(model_all, path)

export_model_typologies_gpkg(model_all, path)
}
\arguments{
\item{model_all}{desc}

\item{path}{desc}
}
\description{
Thes functions write the model object as csv files.
}
