% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exteral_datasets_list}
\alias{exteral_datasets_list}
\title{External Data List}
\format{The data itself is simply a \code{NULL} value.

Its purpose is to provide some documentation of the datasets that are not directly
included in the package.

The following external datasets are accessed by workflows in this package:

\describe{

\item{\code{• kc_boundary}:}{A polygon of King County, WA}
\item{\code{• waterbodies}:}{Multipolygons of the major waterbodies in King County, WA (\href{https://osf.io/9fp6q/}{osf.io/9fp6q})}

}}
\usage{
exteral_datasets_list
}
\description{
This is a list of datasets that are included in the package workflow but
are not stored in the \code{data/} directory.

The primary reason that a dataset is included in this list is that it exceeds Github's
50 MB file size limit (see \href{https://help.github.com/articles/conditions-for-large-files/}{here} for details).
}
\seealso{
\code{\link{plan_external_data}}
}
\keyword{datasets}
