% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factfinder-variables.R
\name{make_factfinder_variables}
\alias{make_factfinder_variables}
\title{Make A Tibble of The Project's US Census Variables from American Factfinder}
\usage{
make_factfinder_variables(factfinder_data, variable_template,
  census_geography_metadata)
}
\arguments{
\item{factfinder_data}{Tibble, desc}

\item{variable_template}{Tibble, desc}

\item{census_geography_metadata}{desc}
}
\value{
a \code{tibble}
}
\description{
Return a \code{tibble} of all of the US Census data variables
that are obtained from the American Factfinder user interface: \link{https://factfinder.census.gov/}
}
