% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{get_data_source_plan}
\alias{get_data_source_plan}
\title{Get the Data Plans}
\usage{
get_data_source_plan()
}
\value{
a \code{drake} plan
}
\description{
Use \code{\link[drake]{drake_plan}} to create the data plan.
}
\examples{

# Print one of the external data plans

get_ext_data_ready_plan()


# Make the plan, load a target, print the target

\dontrun{

plans <- bind_plans(get_ext_data_download_plan(), get_ext_data_ready_plan())

make(plans)

loadd(kc_boundary)

print(kc_boundary)
}
}
