% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-demo-change-indicators.R
\name{make_demo_change_indicators}
\alias{make_demo_change_indicators}
\title{Make The Demographic Change Indicators}
\usage{
make_demo_change_indicators(acs_indicators)
}
\arguments{
\item{acs_indicators}{Tibble, Temporary description.}
}
\value{
a \code{tibble}
}
\description{
Temporary description
}
