% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-indicators-change-in-comparison.R
\name{make_indicators_change_in_comparison}
\alias{make_indicators_change_in_comparison}
\title{Make The Change in Comparison Indicators}
\usage{
make_indicators_change_in_comparison(indicators_comparison,
  change_dategroupid_long, indicator_value_template)
}
\arguments{
\item{change_dategroupid_long}{desc}

\item{indicator_value_template}{desc}

\item{indicators_by_dimension}{desc}
}
\value{
a \code{tibble}
}
\description{
Description
}
