% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-indicators-comparison.R
\name{make_indicators_comparison}
\alias{make_indicators_comparison}
\title{Make The Comparison Indicators}
\usage{
make_indicators_comparison(indicators_by_topic, model_table_production,
  indicator_type_template)
}
\arguments{
\item{indicators_by_topic}{desc}

\item{model_table_production}{desc}

\item{indicator_type_template}{desc}
}
\value{
a \code{tibble}
}
\description{
Description
}
