% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-indicators-comparison-of-change.R
\name{make_indicators_comparison_of_change}
\alias{make_indicators_comparison_of_change}
\title{Make The Change in Comparison Indicators}
\usage{
make_indicators_comparison_of_change(indicators_by_topic, change_endyears,
  indicator_type_template)
}
\arguments{
\item{indicators_by_topic}{desc}

\item{change_endyears}{desc}

\item{indicator_type_template}{desc}
}
\value{
a \code{tibble}
}
\description{
Description
}
