% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-parcel-info-2010.R
\name{make_parcel_info_2010}
\alias{make_parcel_info_2010}
\title{Make The General Tax Data For King County Parcels (2010)}
\usage{
make_parcel_info_2010(dl_parcel_data)
}
\value{
a \code{tibble}
}
\description{
Make the general tax data for King County parcels
(Tax Year 2010, Calendar Year 2009)
}
