% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-parcel-lut-2005.R
\name{make_parcel_lut_2005}
\alias{make_parcel_lut_2005}
\title{Make The Lookup Table for King County Tax Data (2005)}
\usage{
make_parcel_lut_2005(dl_parcel_data)
}
\value{
a \code{tibble}
}
\description{
Make the lookup table for King County parcel tax data (2005).
}
