% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel-lut-2018.R
\name{make_parcel_lut_2018}
\alias{make_parcel_lut_2018}
\title{Make The Lookup Table for King County Tax Data (2018)}
\usage{
make_parcel_lut_2018(zip_path, file_path)
}
\arguments{
\item{zip_path}{Character, the file path of the archive file}

\item{file_path}{Character, the file path of the file that is to be extracted}
}
\value{
a \code{tibble}
}
\description{
Make the lookup table for King County parcel tax data (2018).
}
