% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-parcel-value.R
\name{make_parcel_value}
\alias{make_parcel_value}
\title{Make The Value History For King County Parcels}
\usage{
make_parcel_value(dl_parcel_data)
}
\arguments{
\item{dl_parcel_data}{NULL, a \code{\link[drake]{target}} whos purpose is only
to make sure that the external parcel data file has been downloaded before
\code{parcel_lut_2005} is created.}
}
\value{
a \code{tibble}
}
\description{
Download the history of Tax Assessor appraisals for
all parcels in King County, WA (1998 - 2017).
}
