% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel-value.R
\name{make_parcel_value}
\alias{make_parcel_value}
\title{Make The Value History For King County Parcels}
\usage{
make_parcel_value(zip_path, file_path)
}
\arguments{
\item{zip_path}{Character, the file path of the archive file}

\item{file_path}{Character, the file path of the file that is to be extracted}
}
\value{
a \code{tibble}
}
\description{
Download the history of Tax Assessor appraisals for
all parcels in King County, WA (1998 - 2017).
}
