% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-present-use-key.R
\name{make_present_use_key}
\alias{make_present_use_key}
\title{Make the Key For Decoding Present Use Categories}
\usage{
make_present_use_key(parcel_lut_2005, parcel_lut_2018)
}
\value{
a \code{tibble}
}
\description{
Make the key for decoding the King County Assessor's
"Present Use" categories.
}
