% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-vulnerability-indicators.R
\name{make_vulnerability_indicators}
\alias{make_vulnerability_indicators}
\title{Make The Vulnerability Indicators}
\usage{
make_vulnerability_indicators(acs_indicators)
}
\arguments{
\item{acs_indicators}{desc}
}
\value{
a \code{tibble}
}
\description{
Make The Vulnerability Indicators
}
