% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-models.R
\name{make_model_pdx18}
\alias{make_model_pdx18}
\alias{make_model_coo16}
\alias{make_model_coo18}
\alias{make_model_coorev18}
\alias{make_model_all}
\title{Make The Models}
\usage{
make_model_pdx18(indicators_wide, census_tracts_2016_trimmed)

make_model_coo16(indicators_wide, census_tracts_2016_trimmed)

make_model_coo18(indicators_wide, census_tracts_2016_trimmed)

make_model_coorev18(indicators_wide, census_tracts_2016_trimmed)

make_model_all(indicators_wide, census_tracts_2016_trimmed, model_pdx18,
  model_coo16, model_coo18, model_coorev18)
}
\arguments{
\item{indicators_wide}{desc}
}
\value{
a \code{tibble}
}
\description{
Temporary description.
}
