% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel-data.R
\name{prepare_parcel_data}
\alias{prepare_parcel_data}
\title{Make A Tibble of The Project's Tax Assessor Variables}
\usage{
prepare_parcel_data(model_table, acs_tables, zip_path)
}
\arguments{
\item{model_table}{Tibble, the \code{model_table} object}

\item{acs_tables}{Tibble, the \code{acs_table} object}

\item{zip_path}{Character, the file path of the archive file}

\item{path}{Character, the path or connection to write to.}

\item{file_path}{Character, the file path of the file that is to be extracted}
}
\value{
a \code{tibble}
}
\description{
Return a \code{tibble} of all of the
}
