% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel-data.R
\name{prepare_parcel_data}
\alias{prepare_parcel_data}
\alias{make_parcel_value}
\alias{make_parcel_sales}
\alias{make_parcel_info_2005}
\alias{make_parcel_info_2010}
\alias{make_parcel_info_2015}
\alias{make_parcel_info_2016}
\alias{make_parcel_info_2017}
\alias{make_parcel_info_2018}
\alias{make_parcel_lut_2005}
\alias{make_parcel_lut_2018}
\alias{make_condo_info_2005}
\alias{make_condo_info_2010}
\alias{make_condo_info_2015}
\alias{make_condo_info_2017}
\alias{make_condo_info_2018}
\alias{make_res_bldg_2005}
\alias{make_res_bldg_2010}
\alias{make_res_bldg_2015}
\alias{make_res_bldg_2017}
\alias{make_res_bldg_2018}
\title{Make A Tibble of The Project's Tax Assessor Variables}
\usage{
prepare_parcel_data(model_table, acs_tables, zip_path)

make_parcel_value(data_template, zip_path, file_path)

make_parcel_sales(data_template, zip_path, file_path)

make_parcel_info_2005(metadata_template, zip_path, file_path)

make_parcel_info_2010(metadata_template, zip_path, file_path)

make_parcel_info_2015(metadata_template, zip_path, file_path)

make_parcel_info_2016(metadata_template, zip_path, file_path)

make_parcel_info_2017(metadata_template, zip_path, file_path)

make_parcel_info_2018(metadata_template, zip_path, file_path)

make_parcel_lut_2005(metadata_template, zip_path, file_path)

make_parcel_lut_2018(metadata_template, zip_path, file_path)

make_condo_info_2005(metadata_template, zip_path, file_path)

make_condo_info_2010(metadata_template, zip_path, file_path)

make_condo_info_2015(metadata_template, zip_path, file_path)

make_condo_info_2017(metadata_template, zip_path, file_path)

make_condo_info_2018(metadata_template, zip_path, file_path)

make_res_bldg_2005(metadata_template, zip_path, file_path)

make_res_bldg_2010(metadata_template, zip_path, file_path)

make_res_bldg_2015(metadata_template, zip_path, file_path)

make_res_bldg_2017(metadata_template, zip_path, file_path)

make_res_bldg_2018(metadata_template, zip_path, file_path)
}
\arguments{
\item{model_table}{Tibble, the \code{model_table} object}

\item{acs_tables}{Tibble, the \code{acs_table} object}

\item{zip_path}{Character, the file path of the archive file}

\item{data_template}{Tibble, desc}

\item{file_path}{Character, the file path of the file that is to be extracted}

\item{metadata_template}{Tibble, desc}

\item{path}{Character, the path or connection to write to.}
}
\value{
a \code{tibble}
}
\description{
Return a \code{tibble} of all of the
}
