% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel-variables.R
\name{make_parcel_sales_variables}
\alias{make_parcel_sales_variables}
\alias{make_parcel_value_variables_part1}
\alias{make_parcel_value_variables_part2}
\alias{make_parcel_value_variables}
\title{Make Parcel Variables}
\usage{
make_parcel_sales_variables(parcel_sales, parcel_all_metadata,
  sales_lut_key_list, sales_criteria, single_family_criteria, condo_criteria,
  cpi, variable_template)

make_parcel_value_variables_part1(parcel_all_metadata, single_family_criteria,
  condo_criteria, cpi, parcel_value)

make_parcel_value_variables_part2(parcel_value_variables_part1)

make_parcel_value_variables(parcel_value_variables_part2, variable_template)
}
\arguments{
\item{parcel_sales}{Tibble, Temporary description}

\item{sales_lut_key_list}{Tibble, Temporary description}

\item{sales_criteria}{Tibble, Temporary description}

\item{single_family_criteria}{Tibble, Temporary description}

\item{condo_criteria}{Tibble, Temporary description}

\item{cpi}{Tibble, Temporary description}

\item{present_use_key}{Tibble, Temporary description}

\item{condo_unit_type_key}{Tibble, Temporary description}

\item{parcel_info_2005}{Tibble, Temporary description.}

\item{parcel_info_2010}{Tibble, Temporary description.}

\item{parcel_info_2018}{Tibble, Temporary description.}

\item{condo_info_2010}{Tibble, Temporary description.}

\item{condo_info_2005}{Tibble, Temporary description.}

\item{condo_info_2018}{Tibble, Temporary description.}

\item{res_bldg_2005}{Tibble, Temporary description}

\item{res_bldg_2010}{Tibble, Temporary description}

\item{res_bldg_2018}{Tibble, Temporary description}
}
\value{
a \code{tibble}
}
\description{
King County Assessor data
}
