% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\docType{data}
\name{plan_external_data}
\alias{plan_external_data}
\title{Drake Plan: External Data}
\format{A \code{\link[drake]{drake_plan}} object of class \code{tbl_df} that builds
the following \code{\link[drake]{target}}s:
\describe{

\item{\code{• item_one}:}{Description}

}}
\usage{
plan_external_data
}
\description{
A \code{\link[drake]{drake_plan}} that loads several external datasets.
}
\examples{

# Print the plan

print("change this")


# Make the plan, load a target, print the target

\dontrun{

make(item_one)

readd(item_one)
}
}
\seealso{
\code{\link{data_exteral_datasets_list}}
}
\keyword{drake,}
\keyword{plan}
