% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-typology-map.R
\name{show_typology_map}
\alias{show_typology_map}
\title{SHow Typology Map}
\usage{
show_typology_map(model_pdx16)
}
\arguments{
\item{model_pdx16}{desc}
}
\value{
a MULTIPOLYGON simple feature (class =  \code{sf})
}
\description{
This is a temporary function.
}
