% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-typology-map.R
\name{show_typology_map}
\alias{show_typology_map}
\title{SHow Typology Map}
\usage{
show_typology_map(model_pdx18, model_coo16, model_coo18,
  census_tracts_2016_trimmed)
}
\arguments{
\item{model_pdx18}{desc}

\item{model_coo16}{desc}

\item{census_tracts_2016_trimmed}{desc}
}
\value{
a MULTIPOLYGON simple feature (class =  \code{sf})
}
\description{
This is a temporary function.
}
